refactor: move InternalHookProvider to provider/_protocols.py
